requirements.txt
Flask==2.2.2
google-cloud-texttospeech==2.10.1
paramiko==2.11.0
Werkzeug==2.2.2
